feat: RecommenderPreloader nice still flickering
